+++ lib/sidekiq/scheduled.rb
@@ -33,13 +33,12 @@ module Sidekiq
                 # going wrong between the time jobs are popped from the scheduled queue and when
                 # they are pushed onto a work queue and losing the jobs.
                 while message = conn.zrangebyscore(sorted_set, '-inf', now, :limit => [0, 1]).first do
-                  msg = Sidekiq.load_json(message).tap { |msg| msg['enqueued_at'] = Time.now.to_f }
 
                   # Pop item off the queue and add it to the work queue. If the job can't be popped from
                   # the queue, it's because another process already popped it so we can move on to the
                   # next one.
                   if conn.zrem(sorted_set, message)
-                    Sidekiq::Client.push(msg)
+                    Sidekiq::Client.push(Sidekiq.load_json(message))
                     logger.debug { "enqueued #{sorted_set}: #{message}" }
                   end
                 end